wrote crossplatform tests for desktop and android
