+++ src/main/java/com/solvd/shop24/gui/android/pages/HomePage.java
@@ -1,0 +1,14 @@
+package com.solvd.shop24.gui.android.pages;
+
+import com.qaprosoft.carina.core.foundation.utils.factory.DeviceType;
+import com.qaprosoft.carina.core.foundation.utils.mobile.IMobileUtils;
+import com.solvd.shop24.gui.common.pages.HomePageBase;
+import org.openqa.selenium.WebDriver;
+
+@DeviceType(pageType = DeviceType.Type.ANDROID_PHONE, parentClass = HomePageBase.class)
+public class HomePage extends HomePageBase implements IMobileUtils {
+
+    public HomePage(WebDriver driver) {
+        super(driver);
+    }
+}
