+++ src/main/java/com/solvd/shop24/gui/android/pages/profile/ProfilePage.java
@@ -1,0 +1,14 @@
+package com.solvd.shop24.gui.android.pages.profile;
+
+import com.qaprosoft.carina.core.foundation.utils.factory.DeviceType;
+import com.solvd.shop24.gui.common.pages.profile.AuthenticationPageBase;
+import com.solvd.shop24.gui.common.pages.profile.ProfilePageBase;
+import org.openqa.selenium.WebDriver;
+
+@DeviceType(pageType = DeviceType.Type.ANDROID_PHONE, parentClass = ProfilePageBase.class)
+public class ProfilePage extends ProfilePageBase {
+
+    public ProfilePage(WebDriver driver) {
+        super(driver);
+    }
+}
