--- src/main/java/com/solvd/shop24/gui/pages/constant/PageLocatorsConstants.java
@@ -1,8 +1,0 @@
-package com.solvd.shop24.gui.pages.constant;
-
-import org.openqa.selenium.By;
-
-public class PageLocatorsConstants {
-    public static final By MESSAGE_SUCCESS_ADDING_BASKET = By.className("pp-added");
-    public static final By IS_NOT_IN_STOCK_MESSAGE = By.xpath("//div[text()='К сожалению, товар закончился на складе']");
-}
