--- src/main/java/com/solvd/shop24/gui/pages/news/ArticlePage.java
@@ -1,30 +1,0 @@
-package com.solvd.shop24.gui.pages.news;
-
-import com.beust.jcommander.Parameter;
-import com.qaprosoft.carina.core.foundation.webdriver.decorator.ElementLoadingStrategy;
-import com.qaprosoft.carina.core.foundation.webdriver.decorator.ExtendedWebElement;
-import com.qaprosoft.carina.core.foundation.webdriver.decorator.PageOpeningStrategy;
-import com.solvd.shop24.gui.pages.HomePage;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.support.FindBy;
-
-public class ArticlePage extends HomePage {
-
-    @FindBy(className = "page-main__title")
-    private ExtendedWebElement title;
-
-    @FindBy(xpath = "//div[@class='news-detail']/div[2]")
-    private ExtendedWebElement description;
-
-    public ArticlePage(WebDriver driver) {
-        super(driver);
-    }
-
-    public String getTitle() {
-        return title.getText();
-    }
-
-    public boolean allElementsPresent() {
-        return allElementsPresent(title, description);
-    }
-}
