--- src/main/java/com/solvd/shop24/gui/pages/profile/ProfilePage.java
@@ -1,35 +1,0 @@
-package com.solvd.shop24.gui.pages.profile;
-
-import com.qaprosoft.carina.core.foundation.webdriver.decorator.ExtendedWebElement;
-import com.solvd.shop24.gui.pages.HomePage;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.support.FindBy;
-
-public class ProfilePage extends HomePage {
-
-    @FindBy(xpath = "//dt[text() = 'Имя:']/following::dd[1]")
-    private ExtendedWebElement nameField;
-
-    @FindBy(xpath = "//dt[text() = 'Мобильный телефон:']/following::dd[1]")
-    private ExtendedWebElement phoneNumberField;
-
-    @FindBy(xpath = "//dt[text() = 'E-mail:']/following::dd[1]")
-    private ExtendedWebElement emailField;
-
-    public ProfilePage(WebDriver driver) {
-        super(driver);
-        setPageURL("/cabinet/settings/");
-    }
-
-    public String getName() {
-        return nameField.getText();
-    }
-
-    public String getEmail() {
-        return emailField.getText();
-    }
-
-    public String getPhoneNumber() {
-        return phoneNumberField.getText();
-    }
-}
