+++ src/main/resources/_config.properties
@@ -2,7 +2,7 @@
 #================= Configuration v2 ==================#
 #=====================================================#
 
-#=================== Global configuration ============#
+##=================== Global configuration ============#
 url=https://24shop.by
 platform=*
 browser=chrome
@@ -23,7 +23,7 @@ STAG.api_url=
 #============== Locale Configurations ================#
 locale=en_US
 language=en_US
-browser_language=en_US
+#browser_language=en_US
 enable_l10n=true
 #=====================================================#
 
@@ -39,12 +39,14 @@ thread_count=-1
 #selenium_host=http://localhost:4723/wd/hub
 #driver_type=mobile
 #platform=android
+#browser=chrome
+#url=https://24shop.by
+#capabilities.deviceType=phone
 #capabilities.automationName=UiAutomator2
 #capabilities.STF_ENABLED=false
 #capabilities.platformName=Android
 #capabilities.udid=emulator-5554
 #capabilities.platform=ANDROID
-#capabilities.app=/Users/yantoniuk/Downloads/carinademoexample.apk
 #capabilities.newCommandTimeout=180
 #=====================================================#
