+++ src/main/resources/_testdata.properties
@@ -6,3 +6,4 @@ account.password=cmn0hd
 account.phoneNumber=+375 44 454-42-81
 account.name=Alex
 account.email=y@mail.ru
+account.authPhoneNumber=444544281
