+++ src/test/java/com/solvd/shop24/NewsTest.java
@@ -2,10 +2,10 @@ package com.solvd.shop24;
 
 import com.qaprosoft.carina.core.foundation.AbstractTest;
 import com.qaprosoft.carina.core.foundation.utils.ownership.MethodOwner;
-import com.solvd.shop24.gui.components.news.ArticleItem;
-import com.solvd.shop24.gui.pages.news.ArticlePage;
-import com.solvd.shop24.gui.pages.HomePage;
-import com.solvd.shop24.gui.pages.news.NewsPage;
+import com.solvd.shop24.gui.common.components.news.ArticleItem;
+import com.solvd.shop24.gui.common.pages.HomePageBase;
+import com.solvd.shop24.gui.common.pages.news.ArticlePageBase;
+import com.solvd.shop24.gui.common.pages.news.NewsPageBase;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 import org.testng.asserts.SoftAssert;
@@ -17,11 +17,11 @@ public class NewsTest extends AbstractTest {
     @Test
     @MethodOwner(owner = "yantoniuk")
     public void testReadNews() {
-        HomePage homePage = new HomePage(getDriver());
+        HomePageBase homePage = initPage(getDriver(), HomePageBase.class);
         homePage.open();
         Assert.assertTrue(homePage.isPageOpened(), "Home page wasn't opened!");
 
-        NewsPage newsPage = homePage.getMenu().openNews();
+        NewsPageBase newsPage = homePage.getMenu().openNews();
         newsPage.assertPageOpened();
         Assert.assertEquals(newsPage.getTitle(), "Новости", "Invalid title of News page!");
         Assert.assertFalse(newsPage.getArticlesList().isEmpty(), "Empty article's list!");
@@ -29,7 +29,7 @@ public class NewsTest extends AbstractTest {
         ArticleItem articleItem = newsPage.getArticle(articleIndex);
         String title = articleItem.getTitle();
 
-        ArticlePage articlePage = articleItem.openArticle();
+        ArticlePageBase articlePage = articleItem.openArticle();
         new SoftAssert().assertTrue(articlePage.allElementsPresent(),
                 "Article page hasn't a description or a title!");
         Assert.assertEquals(articlePage.getTitle(), title, "Invalid article was opened!");
